refactor(ff-probe): split probe_inner.rs into focused submodules
